refactor(usager): schema organisme et regex dans utils
