+++ packages/frontend-usagers/src/components/operateur/personne-morale.vue
@@ -203,38 +203,11 @@ const headers = [
   },
 ];
 
-const numTelephoneRegex = /^(\+33|0|0033)[1-9][0-9]{8}$/i;
-const siretRegex = /^[0-9]{14}$/;
 const randomId = ref(random.getRandomId());
 const personneMorale = ref();
 const operateurDejaExistant = ref();
 
-// Schéma et données liées à une personne morale
-const schema = {
-  siret: yup
-    .string()
-    .test(
-      "siret",
-      "Le numéro SIRET doit faire exactement 14 chiffres, sans espace",
-      (siret) => siretRegex.test(siret),
-    )
-    .required(),
-  email: yup
-    .string()
-    .email("le format de l'email n'est pas valide")
-    .required("L'email de contact est obligatoire"),
-  telephoneEP: yup
-    .string()
-    .test(
-      "telephone",
-      "Format de numéro de téléphone invalide",
-      (telephoneEP) => numTelephoneRegex.test(telephoneEP),
-    )
-    .required("Le numéro de téléphone de l'établissement est obligatoire"),
-  representantsLegaux: yup.array().min(1, "pas bien").required(),
-  etablissements: yup.array().min(1, "pas bien").required(),
-};
-const validationSchema = yup.object({ ...schema });
+const validationSchema = yup.object({ ...organisme.schema.personneMorale });
 
 const initialValues = {
   siret: props.initData.siret,
