+++ packages/frontend-usagers/src/components/personne.vue
@@ -132,80 +132,12 @@ const props = defineProps({
 
 const emit = defineEmits(["valid", "update:personne"]);
 
-const emailRegex = /^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/;
-const numTelephoneRegex = /^(\+33|0|0033)[1-9][0-9]{8}$/i;
-const acceptedCharsRegex =
-  /^([AÀÂBCÇDEÉÈÊËFGHIÎÏJKLMNOÔPQRSTUÙÛÜVWXYŸZÆŒ\- ']+)$/i;
-const spaceFollowingDashRegex = /( -)|(- )/i;
-const doubleSpacesRegex = / {2}/i;
-const tripleDashRegex = /-{3}/i;
-const doubleDashRegex = /-{2}/i;
-
 const validationSchema = computed(() =>
   yup.object({
-    nom: yup
-      .string()
-      .test("acceptedChars", "Caractères non acceptés détectés", (nom) =>
-        acceptedCharsRegex.test(nom),
-      )
-      .test(
-        "doubleSpaces",
-        "Le nom ne peut contenir deux espaces successifs",
-        (nom) => !doubleSpacesRegex.test(nom),
-      )
-      .test(
-        "spaceFollowingDash",
-        "Le nom ne peut contenir d'espace suivant un tiret",
-        (nom) => !spaceFollowingDashRegex.test(nom),
-      )
-      .test(
-        "tripleDash",
-        "Le nom ne peut contenir trois tirets consécutifs",
-        (nom) => !tripleDashRegex.test(nom),
-      )
-      .required(),
-    prenom: yup
-      .string()
-      .test("acceptedChars", "Caractères non acceptés détectés", (prenom) =>
-        acceptedCharsRegex.test(prenom),
-      )
-      .test(
-        "doubleSpaces",
-        "Le prénom ne peut contenir deux espaces successifs",
-        (prenom) => !doubleSpacesRegex.test(prenom),
-      )
-      .test(
-        "spaceFollowingDash",
-        "Le prénom ne peut contenir d'espace suivant un tiret",
-        (prenom) => !spaceFollowingDashRegex.test(prenom),
-      )
-      .test(
-        "doubleDash",
-        "Le prénom ne peut contenir deux tirets consécutifs",
-        (prenom) => !doubleDashRegex.test(prenom),
-      )
-      .required(),
-    fonction: yup.string().required(),
-    ...(props.showTelephone && {
-      telephone: yup
-        .string()
-        .test(
-          "telephone",
-          "Format de numéro de téléphone invalide",
-          (telephone) => numTelephoneRegex.test(telephone),
-        )
-        .required(),
-    }),
-    ...(props.showEmail && {
-      email: yup
-        .string()
-        .test("email", "l'email n'est pas au format attendu", (email) =>
-          emailRegex.test(email),
-        )
-        .required(),
-    }),
-    ...(props.showAdresse && {
-      adresse: yup.object().required(),
+    ...personne.schema({
+      showAdresse: props.showAdresse,
+      showTelephone: props.showTelephone,
+      showEmail: props.showEmail,
     }),
   }),
 );
