+++ packages/frontend-usagers/src/components/protocole-sanitaire.vue
@@ -47,7 +47,7 @@
             v-model="constitutionEquipe"
             name="constitutionEquipe"
             legend="L’équipe comprend-elle ?"
-            :options="constitutionEquipeOptions"
+            :options="protocoleSanitaire.constitutionEquipeOptions"
             :small="true"
             :required="true"
           />
@@ -96,7 +96,9 @@
             v-model="responsableAdministrationMedicament"
             name="responsableAdministrationMedicament"
             legend="Personne désignée en charge de la distribution et de l’administration des médicaments et de l’enregistrement de l’administration"
-            :options="responsableAdministrationMedicamentOptions"
+            :options="
+              protocoleSanitaire.responsableAdministrationMedicamentOptions
+            "
             :small="true"
             :required="true"
           />
@@ -104,7 +106,7 @@
       </div>
     </fieldset>
     <fieldset
-      v-if="responsableAdministrationMedicament.length > 0"
+      v-if="responsableAdministrationMedicament?.length > 0"
       class="fr-fieldset"
     >
       <div class="fr-fieldset__element">
@@ -252,7 +254,7 @@
             legend="Méthode retenue pour la préparation des piluliers"
             :required="true"
             :model-value="preparationPilluliers"
-            :options="preparationPilluliersOptions"
+            :options="protocoleSanitaire.preparationPilluliersOptions"
             :is-valid="preparationPilluliersMeta"
             :inline="true"
             :error-message="preparationPilluliersErrorMessage"
@@ -539,227 +541,57 @@ import { useField, useForm } from "vee-validate";
 import * as yup from "yup";
 
 const props = defineProps({
-  initData: { type: Object, default: null, required: true },
+  initData: { type: Object, required: true },
 });
 const emit = defineEmits(["previous", "next", "update"]);
 
 const log = logger("components/protocole-sanitaire");
 
-const constitutionEquipeOptions = [
-  {
-    label:
-      "Une personne formée aux gestes et soins d’urgence (PSC 1, SST, AFGSU 1, AFGSU 2, AFSGSU)",
-    id: "personne_formee",
-    name: "personne_formee",
-  },
-  {
-    label: "Un(e) infirmier(e)",
-    id: "infirmier",
-    name: "infirmier",
-  },
-  {
-    label: "Un(e) aide soignant(e)",
-    id: "aide_soignant",
-    name: "aide_soignant",
-  },
-];
-
-const responsableAdministrationMedicamentOptions = [
-  {
-    label: "Responsable du séjour",
-    id: "responsable_sejour",
-    name: "responsable_sejour",
-  },
-  {
-    label: "Accompagnant(s)",
-    id: "accompagnant",
-    name: "accompagnant",
-  },
-  {
-    label: "Professionnel de santé",
-    id: "professionnel_sante",
-    name: "professionnel_sante",
-  },
-];
-
-const preparationPilluliersOptions = [
-  { label: "Aucune méthode", id: "aucune", value: "aucune" },
-  {
-    label:
-      "Piluliers préparés préalablement au séjour par le vacancier, sa famille, le représentant légal, l’établissement de résidence habituelle, le médecin",
-    id: "prepares_prealablement",
-    value: "prepares_prealablement",
-  },
-  {
-    label: "Piluliers préparés durant le séjour",
-    id: "au_fur_et_a_mesure",
-    value: "au_fur_et_a_mesure",
-  },
-];
-
-const schemaInfosSanitaires = {
-  dispositionsSpecifiques: yup.boolean().required(),
-  precisionDispositionsSpecifiques: yup
-    .string()
-    .when("dispositionsSpecifiques", {
-      is: (val) => !!val,
-      then: (precision) =>
-        precision
-          .min(5, "Vous devez préciser votre réponse précédente")
-          .required(),
-      otherwise: (precision) => precision.nullable(),
-    }),
-  constitutionEquipe: yup.array().required(),
-  precisionConstitutionEquipe: yup.string().when("constitutionEquipe", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  troussePharmacie: yup.boolean().required(),
-  responsableAdministrationMedicament: yup.array().required(),
-  stockageMedicamentSecurise: yup.boolean().required(),
-  precisionStockageMedicamentSecurise: yup
-    .string()
-    .when("stockageMedicamentSecurise", {
-      is: (val) => !!val,
-      then: (precision) =>
-        precision
-          .min(5, "Vous devez préciser votre réponse précédente")
-          .required(),
-      otherwise: (precision) => precision.nullable(),
-    }),
-  conservationMedicamentThermosensible: yup.boolean().required(),
-  precisionConservationMedicament: yup.string().when("conservationMedicament", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  individualisationMedicaments: yup.boolean().required(),
-  precisionIndividualisationMedicaments: yup
-    .string()
-    .when("individualisationMedicaments", {
-      is: (val) => !!val,
-      then: (precision) =>
-        precision
-          .min(5, "Vous devez préciser votre réponse précédente")
-          .required(),
-      otherwise: (precision) => precision.nullable(),
-    }),
-  preparationPilluliers: yup.string().required(),
-  precisionPreparationPilluliers: yup.string().when("preparationPilluliers", {
-    is: (preparationPilluliers) =>
-      preparationPilluliers === "prepares_prealablement" ||
-      preparationPilluliers === "au_fur_et_a_mesure",
-
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  prescriptionMedicaleJointe: yup.boolean().required(),
-  protocoleModificationTraitement: yup.boolean().required(),
-  precisionProtocoleModificationTraitement: yup
-    .string()
-    .when("protocoleModificationTraitement", {
-      is: (val) => !!val,
-      then: (precision) =>
-        precision
-          .min(5, "Vous devez préciser votre réponse précédente")
-          .required(),
-      otherwise: (precision) => precision.nullable(),
-    }),
-  ficheSuiviMedicaments: yup.string().required(),
-  protocoleEvacuation: yup.boolean().required(),
-  precisionProtocoleEvacuation: yup.string().when("protocoleEvacuation", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  protocoleAccident: yup.boolean().required(),
-  precisionProtocoleAccident: yup.string().when("protocoleAccident", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  protocoleReorientation: yup.boolean().required(),
-  precisionProtocoleReorientation: yup.string().when("protocoleReorientation", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  //
-  protocoleCanicule: yup.boolean().required(),
-  precisionProtocoleCanicule: yup.string().when("protocoleCanicule", {
-    is: (val) => !!val,
-    then: (precision) =>
-      precision
-        .min(5, "Vous devez préciser votre réponse précédente")
-        .required(),
-    otherwise: (precision) => precision.nullable(),
-  }),
-  gestionBudgetPersonnel: yup.string().required(),
-};
+const schemaInfosSanitaires = { ...protocoleSanitaire.schema };
 
 const validationSchema = yup.object({
   ...schemaInfosSanitaires,
 });
 
 const initialValues = {
-  dispositionsSpecifiques: props.initData?.dispositionsSpecifiques,
+  dispositionsSpecifiques: props.initData.dispositionsSpecifiques,
   precisionDispositionsSpecifiques:
-    props.initData?.precisionDispositionsSpecifiques,
-  constitutionEquipe: props.initData?.constitutionEquipe ?? [],
-  precisionConstitutionEquipe: props.initData?.precisionConstitutionEquipe,
-  troussePharmacie: props.initData?.troussePharmacie,
+    props.initData.precisionDispositionsSpecifiques,
+  constitutionEquipe: props.initData.constitutionEquipe ?? [],
+  precisionConstitutionEquipe: props.initData.precisionConstitutionEquipe,
+  troussePharmacie: props.initData.troussePharmacie,
   responsableAdministrationMedicament:
-    props.initData?.responsableAdministrationMedicament ?? [],
+    props.initData.responsableAdministrationMedicament ?? [],
   precisionResponsableAdministrationMedicament:
-    props.initData?.precisionResponsableAdministrationMedicament,
-  stockageMedicamentSecurise: props.initData?.stockageMedicamentSecurise,
+    props.initData.precisionResponsableAdministrationMedicament,
+  stockageMedicamentSecurise: props.initData.stockageMedicamentSecurise,
   precisionStockageMedicamentSecurise:
-    props.initData?.precisionStockageMedicamentSecurise,
+    props.initData.precisionStockageMedicamentSecurise,
   conservationMedicamentThermosensible:
-    props.initData?.conservationMedicamentThermosensible,
+    props.initData.conservationMedicamentThermosensible,
   precisionConservationMedicament:
-    props.initData?.precisionConservationMedicament,
-  individualisationMedicaments: props.initData?.individualisationMedicaments,
+    props.initData.precisionConservationMedicament,
+  individualisationMedicaments: props.initData.individualisationMedicaments,
   precisionIndividualisationMedicaments:
-    props.initData?.precisionIndividualisationMedicaments,
-  preparationPilluliers: props.initData?.preparationPilluliers,
-  precisionPreparationPilluliers:
-    props.initData?.precisionPreparationPilluliers,
-  prescriptionMedicaleJointe: props.initData?.prescriptionMedicaleJointe,
+    props.initData.precisionIndividualisationMedicaments,
+  preparationPilluliers: props.initData.preparationPilluliers,
+  precisionPreparationPilluliers: props.initData.precisionPreparationPilluliers,
+  prescriptionMedicaleJointe: props.initData.prescriptionMedicaleJointe,
   protocoleModificationTraitement:
-    props.initData?.protocoleModificationTraitement,
+    props.initData.protocoleModificationTraitement,
   precisionProtocoleModificationTraitement:
-    props.initData?.precisionProtocoleModificationTraitement,
-  ficheSuiviMedicaments: props.initData?.ficheSuiviMedicaments,
-  protocoleEvacuation: props.initData?.protocoleEvacuation,
-  precisionProtocoleEvacuation: props.initData?.precisionProtocoleEvacuation,
-  protocoleAccident: props.initData?.protocoleAccident,
-  precisionProtocoleAccident: props.initData?.precisionProtocoleAccident,
-  protocoleReorientation: props.initData?.protocoleReorientation,
+    props.initData.precisionProtocoleModificationTraitement,
+  ficheSuiviMedicaments: props.initData.ficheSuiviMedicaments,
+  protocoleEvacuation: props.initData.protocoleEvacuation,
+  precisionProtocoleEvacuation: props.initData.precisionProtocoleEvacuation,
+  protocoleAccident: props.initData.protocoleAccident,
+  precisionProtocoleAccident: props.initData.precisionProtocoleAccident,
+  protocoleReorientation: props.initData.protocoleReorientation,
   precisionProtocoleReorientation:
-    props.initData?.precisionProtocoleReorientation,
-  protocoleCanicule: props.initData?.protocoleCanicule,
-  precisionProtocoleCanicule: props.initData?.precisionProtocoleCanicule,
-  gestionBudgetPersonnel: props.initData?.gestionBudgetPersonnel,
+    props.initData.precisionProtocoleReorientation,
+  protocoleCanicule: props.initData.protocoleCanicule,
+  precisionProtocoleCanicule: props.initData.precisionProtocoleCanicule,
+  gestionBudgetPersonnel: props.initData.gestionBudgetPersonnel,
 };
 
 const { meta, values } = useForm({
