+++ packages/frontend-usagers/src/components/protocole-transport-read-only.vue
@@ -57,7 +57,7 @@
 const log = logger("components/protocole-transport-read-only");
 
 const props = defineProps({
-  initData: { type: Object, default: null, required: true },
+  initData: { type: Object, required: true },
 });
 </script>
