+++ packages/frontend-usagers/src/utils/regex.js
@@ -1,0 +1,31 @@
+const numTelephoneRegex = /^(\+33|0|0033)[1-9][0-9]{8}$/i;
+const emailRegex = /^[\w-.]+@([\w-]+\.)+[\w-]{2,4}$/;
+const siretRegex = /^[0-9]{14}$/;
+const acceptedCharsRegex =
+  /^([AÀÂBCÇDEÉÈÊËFGHIÎÏJKLMNOÔPQRSTUÙÛÜVWXYŸZÆŒ\- ']+)$/i;
+const spaceFollowingDashRegex = /( -)|(- )/i;
+const doubleSpacesRegex = / {2}/i;
+const tripleDashRegex = /-{3}/i;
+const doubleDashRegex = /-{2}/i;
+const pwdRegex =
+  /^(?=.*[a-z])(?=.*[A-Z])(?=.*[0-9])(?=.*[!@#$%^&*)(<>,~;])(?=.{12,})/;
+const pwdRegexSpecial = /^(?=.*[!@#$%^&*)(<>,~;])/;
+const pwdRegexNumber = /^(?=.*[0-9])/;
+const pwdRegexMaj = /^(?=.*[A-Z])/;
+const pwdRegexMin = /^(?=.*[a-z])/;
+
+export default {
+  numTelephoneRegex,
+  emailRegex,
+  siretRegex,
+  acceptedCharsRegex,
+  spaceFollowingDashRegex,
+  doubleSpacesRegex,
+  tripleDashRegex,
+  doubleDashRegex,
+  pwdRegex,
+  pwdRegexSpecial,
+  pwdRegexNumber,
+  pwdRegexMaj,
+  pwdRegexMin,
+};
